Move _widgetFromAction and buildMaterialActionButton to AppScaffoldAction class
