+++ lib/pages/classes/view/classes_page.dart
@@ -4,6 +4,8 @@ import 'package:acs_upb_mobile/pages/class_feedback/view/class_feedback_checklis
 import 'package:acs_upb_mobile/pages/classes/model/class.dart';
 import 'package:acs_upb_mobile/pages/classes/service/class_provider.dart';
 import 'package:acs_upb_mobile/pages/classes/view/class_view.dart';
+import 'package:acs_upb_mobile/resources/remote_config.dart';
+import 'package:acs_upb_mobile/widgets/app_scaffold_action.dart';
 import 'package:acs_upb_mobile/widgets/class_icon.dart';
 import 'package:acs_upb_mobile/widgets/error_page.dart';
 import 'package:acs_upb_mobile/widgets/icon_text.dart';
@@ -12,7 +14,6 @@ import 'package:acs_upb_mobile/widgets/spoiler.dart';
 import 'package:flutter/material.dart';
 import 'package:pedantic/pedantic.dart';
 import 'package:provider/provider.dart';
-import 'package:acs_upb_mobile/resources/remote_config.dart';
 
 class ClassesPage extends StatefulWidget {
   const ClassesPage({Key key}) : super(key: key);
