+++ lib/pages/people/view/people_page.dart
@@ -2,6 +2,7 @@ import 'package:acs_upb_mobile/generated/l10n.dart';
 import 'package:acs_upb_mobile/pages/people/model/person.dart';
 import 'package:acs_upb_mobile/pages/people/service/person_provider.dart';
 import 'package:acs_upb_mobile/pages/people/view/person_view.dart';
+import 'package:acs_upb_mobile/widgets/app_scaffold_action.dart';
 import 'package:acs_upb_mobile/widgets/autocomplete.dart';
 import 'package:acs_upb_mobile/widgets/scaffold.dart';
 import 'package:acs_upb_mobile/widgets/search_bar.dart';
@@ -39,12 +40,11 @@ class _PeoplePageState extends State<PeoplePage> {
     return AppScaffold(
       actions: [
         AppScaffoldAction(
-          icon: Icons.search_outlined,
-          onPressed: () {
-            setState(() => searchClosed = !searchClosed);
-          },
-          tooltip: S.current.actionSearch
-        )
+            icon: Icons.search_outlined,
+            onPressed: () {
+              setState(() => searchClosed = !searchClosed);
+            },
+            tooltip: S.current.actionSearch)
       ],
       title: Text(S.current.navigationPeople),
       body: FutureBuilder(
