+++ lib/widgets/scaffold.dart
@@ -8,40 +8,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import 'action_sidebar.dart';
-
-class AppScaffoldAction {
-  AppScaffoldAction({
-    this.icon,
-    this.route,
-    this.onPressed,
-    this.tooltip,
-    this.text,
-    this.items,
-    bool disabled,
-  }) : disabled = disabled ?? false;
-
-  // Icon for the action button
-  final IconData icon;
-
-  // Route to push when the action button is pressed
-  final String route;
-
-  // Action that happens when the button is pressed. This overrides [route].
-  final void Function() onPressed;
-
-  // Action button tooltip text
-  final String tooltip;
-
-  // Action text. This overrides [icon].
-  final String text;
-
-  // Option-action map that should be specified if a popup menu is needed. It
-  // overrides [route].
-  final Map<String, void Function()> items;
-
-  // Whether the icon color should be disabled.
-  final bool disabled;
-}
+import 'app_scaffold_action.dart';
 
 class AppScaffold extends StatelessWidget {
   AppScaffold({
@@ -63,58 +30,6 @@ class AppScaffold extends StatelessWidget {
   final bool needsToBeAuthenticated;
   final bool onWeb;
   final double maxBodyWidth;
-
-  Widget _widgetFromAction(AppScaffoldAction action,
-      {@required bool enableContent, @required BuildContext context}) {
-    if (action == null) {
-      return null;
-    }
-
-    final void Function() onPressed =
-        !enableContent || (action?.onPressed == null && action?.route == null)
-            ? null
-            : action?.onPressed ??
-                () => Navigator.pushNamed(context, action?.route);
-
-    final icon = action.disabled
-        ? Icon(action.icon ?? Icons.menu_outlined,
-            color: Theme.of(context).disabledColor)
-        : Icon(action.icon);
-
-    return action?.items != null
-        ? PopupMenuButton<String>(
-            icon: icon,
-            tooltip: action.tooltip ?? action.text,
-            onSelected: (selected) => action.items[selected](),
-            itemBuilder: (BuildContext context) {
-              return action.items.keys
-                  .map((option) => PopupMenuItem(
-                        value: option,
-                        child: Text(option),
-                      ))
-                  .toList();
-            },
-          )
-        : Tooltip(
-            message: action?.tooltip ?? action?.text ?? '',
-            child: action?.text != null
-                ? ButtonTheme(
-                    minWidth: 10,
-                    child: TextButton(
-                      child: Text(
-                        action.text,
-                        style: const TextStyle().apply(
-                            color: Theme.of(context).primaryIconTheme.color),
-                      ),
-                      onPressed: onPressed,
-                    ),
-                  )
-                : IconButton(
-                    icon: icon,
-                    onPressed: onPressed,
-                  ),
-          );
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -151,11 +66,13 @@ class AppScaffold extends StatelessWidget {
                   title: title,
                   centerTitle: true,
                   toolbarOpacity: 0.8,
-                  leading: _widgetFromAction(leading,
+                  leading: leading?.toWidget(
                       enableContent: enableContent, context: context),
                   actions: actions
-                      .map((action) => _widgetFromAction(action,
-                          enableContent: enableContent, context: context))
+                      .map(
+                        (action) => action?.toWidget(
+                            enableContent: enableContent, context: context),
+                      )
                       .toList(),
                 ),
               )
@@ -185,7 +102,7 @@ class AppScaffold extends StatelessWidget {
                       child: body)),
               if (actionsList.isNotEmpty)
                 ActionSideBar(actionsList
-                    .map((e) => _widgetFromAction(e,
+                    .map((e) => e?.toMaterialActionButton(
                         enableContent: enableContent, context: context))
                     .where((element) => element != null)
                     .toList())
@@ -194,5 +111,4 @@ class AppScaffold extends StatelessWidget {
             ],
           );
   }
-
 }
